pyproject.toml
[project]
name = "aivi-linebot"
version = "0.1.0"
description = "AI 新聞推播 LINE Bot - 自動爬取新聞並透過 LINE 推播"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "line-bot-sdk>=3.0.0",
    "httpx>=0.27.0",
    "selectolax>=0.3.0",
    "flask>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=term-missing"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "*/__init__.py"]
